de-monolith Stage 6a: relocate the lean corridor + trajectory substrate to kirra-core
Extract the dependency-light geometry/perception substrate out of the heavy
kirra-ros2-adapter into kirra-core, so the shared lane map (kirra-map, Stage 6b)
and the planner can build on it without pulling the adapter's tokio/dashmap/r2r
tree. Pure move + re-export — every existing path keeps the SAME type.

Moved to kirra-core (verbatim — fields, variants, derives, trait, and impls all
byte-identical; only doc comments / file headers differ):
- `kirra_core::corridor` <- `Point`, `CorridorSource`, `MockCorridorSource` (+ the
  3 corridor unit tests) from the adapter's `corridor/mod.rs`.
- `kirra_core::trajectory` <- `Pose`, `TrajectoryPoint`, `TrajectoryVerdict`,
  `PerceivedObject` from the adapter's `state.rs`.

Stays in the adapter (correctly heavy / coupled), now consuming the moved types via
re-export:
- `AdaptorState` (DashMap + PostureTracker concurrency model), `AcceptedTrajectory`
  (the per-asset slot record + its state-machine tests), `EgoOdom`,
  `IncomingTrajectory`, and all the AdaptorState tests.
- The feature-gated Lanelet2 cxx bridge (`lanelet2` / `lanelet2_bridge`, C++ linkage)
  — `Lanelet2CorridorSource` now impls the relocated `kirra_core` trait for its local
  type (a coherent foreign-trait-for-local-type impl).

The adapter's `corridor` and `state` modules become thin re-export surfaces
(`pub use kirra_core::corridor::*` / `::trajectory::*`), so every existing
`kirra_ros2_adapter::corridor::{Point, CorridorSource, MockCorridorSource}` and
`::state::{Pose, TrajectoryPoint, TrajectoryVerdict, PerceivedObject}` path — used
across the planner, validation.rs, node.rs, and the tests — is unchanged. The adapter
gains a direct `kirra-core` dependency; it KEEPS `kirra-runtime-sdk` (validation.rs
still needs perception_monitor + the containment/kinematics kernel — a later step if
the adapter is to fully shed the heavy dep).

This unblocks Stage 6b: lifting the planner's `lanemap.rs` (LaneGraph / right-of-way /
routing) into a standalone `kirra-map` crate on this lean substrate — the shared
lane map originally requested.

Verification: root workspace build + tests green (kirra-core 75 incl. the 3 relocated
corridor tests; adapter 42 + 22 validation + integration suites green via re-export;
planner 102 + integration green, consuming the substrate transparently); CI-equivalent
`cargo clippy --workspace -- -D warnings` clean; parko workspace logic green
(parko-core 204, parko-kirra 148, parko-ros2 168); no `#[path]` hazards; the four moved
types proven field/variant/derive-identical and the corridor code proven identical by
diff.
